version-9/requirements.txt
Flask
grpcio
grpcio-tools
opencv-python
mediapipe
numpy
tensorflow
scikit-learn
